% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_id_race.R
\name{give_id_race}
\alias{give_id_race}
\title{Extract race IDs from the FFCK website for given years}
\usage{
give_id_race(years_tab)
}
\arguments{
\item{years_tab}{A vector of integers representing the years to scrape.}
}
\value{
A data frame with columns \code{annee}, \code{ID_competition}, and \code{ID_course} representing the year, competition ID, and race ID respectively.
}
\description{
This function scrapes the FFCK website to extract competition and race IDs for the specified years.
}
\examples{
give_id_race(c(2001, 2002))
}
