% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_years.R
\name{give_years}
\alias{give_years}
\title{Give the years present on the FFCK website}
\usage{
give_years()
}
\value{
A vector of integers representing the years mentioned on the FFCK website.
}
\description{
This function scrapes the FFCK website for the years listed their database.
}
\examples{
give_years()
}
