% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maj_data.R
\name{maj_data}
\alias{maj_data}
\title{Update Data Function}
\usage{
maj_data(link_database_csv, link_error_csv)
}
\arguments{
\item{link_database_csv}{A character string containing the file path to the CSV
file that contains the database of courses.}

\item{link_error_csv}{A character string containing the file path to the CSV
file for logging errors encountered during data retrieval.}
}
\value{
A list of years printed to the console. The function does not return any
value (invisible NULL).
}
\description{
This function retrieves a list of years using the \code{give_years} function and prints
the result. It is designed to be a part of a larger data management workflow where
updating or processing data for specific years is necessary. It checks for existing
courses in a database and scraps data for new courses if necessary.
}
