% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_course.R
\name{scrap_course}
\alias{scrap_course}
\title{Scrap Course Results from FFCK}
\usage{
scrap_course(ID_course, ID_competition, year, isN1 = FALSE)
}
\arguments{
\item{ID_course}{A numeric or character ID representing the course to be scraped.}

\item{ID_competition}{A numeric or character ID representing the competition to which the course belongs.}

\item{year}{A numeric year for the competition.}

\item{isN1}{A logical value indicating whether the competition is a national event (N1). Default is FALSE.}
}
\value{
A data frame containing the scraped course results, including competitor information
and relevant statistics.
}
\description{
This function scrapes the course results from the French canoeing federation website
using the specified course ID, competition ID, year, and whether it's a national event (N1).
}
