% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_course.R
\name{verif_course}
\alias{verif_course}
\title{Verify Course import}
\usage{
verif_course(df_test)
}
\arguments{
\item{df_test}{A data frame containing course names in the column 'Nom'.}
}
\value{
A logical value indicating whether the number of course names
that are purely numeric exceeds 10\% of the total number of names.
}
\description{
This function checks if the proportion of course names consisting solely of numbers
exceeds a specified threshold (10\% of the total number of entries).
}
